Django-Middleware-0x03/requirements.txt
# Core dependencies
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.2
django-environ>=0.11.2
drf-nested-routers>=0.93.4



